// src/app/master/models/custdetconsignee.ts
import { GlobalVariables } from '../../core/models/globalvariables';

export class CustdetConsignee {

    det_pkid: string;

    det_cust_id: string;
    rec_mode: string;

    det_branch_id: string;
    det_branch_code: string;
    det_branch_name: string;

    det_sman_id: string;
    det_sman_name: string;

    det_consignee_id: string;
    det_consignee_code: string;
    det_consignee_name: string;


    _globalvariables: GlobalVariables;



}